feat(bundletrial): rename folder for bundle trial (#3123)
