+++ apps/frontend/src/components/service/trial-instances/xtm-platform-trial/manage-trial/TrialUserRolePanelFields.tsx
